+++ cloud/filestore/private/api/protos/tablet.proto
@@ -121,11 +121,6 @@ message TStorageStats
     uint64 LastCollectCommitId = 112;
     uint64 DeletedFreshBytesCount = 113;
 
-    uint64 NodesOpenForWritingBySingleSession = 114;
-    uint64 NodesOpenForWritingByMultipleSessions = 115; 
-    uint64 NodesOpenForReadingBySingleSession = 116;
-    uint64 NodesOpenForReadingByMultipleSessions = 117;
-
     // channel stats
     uint64 TabletChannelCount = 1000;
     uint64 ConfigChannelCount = 1001;
